Remove activity for second fragment and move into directory
