+++ app/src/main/java/io/homeassistant/companion/android/PresenterComponent.kt
@@ -11,8 +11,7 @@ import io.homeassistant.companion.android.onboarding.manual.ManualSetupFragment
 import io.homeassistant.companion.android.settings.SettingsActivity
 import io.homeassistant.companion.android.settings.SettingsFragment
 import io.homeassistant.companion.android.settings.ssid.SsidDialogFragment
-import io.homeassistant.companion.android.shortcuts.ShortcutsActivity
-import io.homeassistant.companion.android.shortcuts.ShortcutsFragment
+import io.homeassistant.companion.android.settings.shortcuts.ShortcutsFragment
 import io.homeassistant.companion.android.webview.WebViewActivity
 
 @Component(dependencies = [AppComponent::class], modules = [PresenterModule::class])
@@ -31,8 +30,6 @@ interface PresenterComponent {
     fun inject(activity: SettingsActivity)
 
     fun inject(fragment: SettingsFragment)
-
-    fun inject(activity: ShortcutsActivity)
 
     fun inject(fragment: ShortcutsFragment)
