+++ app/src/main/java/io/homeassistant/companion/android/settings/shortcuts/ShortcutsPresenter.kt
@@ -1,11 +1,7 @@
-package io.homeassistant.companion.android.shortcuts
+package io.homeassistant.companion.android.settings.shortcuts
 
 import io.homeassistant.companion.android.domain.integration.Panel
 
 interface ShortcutsPresenter {
-    fun onCreate()
-
-    fun onFinish()
-
     fun getPanels(): Array<Panel>
 }
