+++ app/src/main/java/io/homeassistant/companion/android/settings/shortcuts/ShortcutsView.kt
@@ -1,0 +1,3 @@
+package io.homeassistant.companion.android.settings.shortcuts
+
+interface ShortcutsView
