--- app/src/main/java/io/homeassistant/companion/android/shortcuts/ShortcutsView.kt
@@ -1,3 +1,0 @@
-package io.homeassistant.companion.android.shortcuts
-
-interface ShortcutsView
